Chore #265 - Const 파일 하나로 통합
