--- Baggle/Baggle/Core/Common/Const/Account.swift
@@ -1,14 +1,0 @@
-//
-//  Account.swift
-//  Baggle
-//
-//  Created by 양수빈 on 2023/09/19.
-//
-
-import Foundation
-
-extension Const {
-    struct Account {
-        static let mail = "kingzzangssang@gmail.com"
-    }
-}
