+++ Baggle/Baggle/Core/Common/Const/Const.swift
@@ -7,4 +7,26 @@
 
 import Foundation
 
-struct Const { }
+// swiftlint:disable all
+struct Const {
+    struct URL {
+        static let privacyPolicy = "https://www.dnd.ac/"
+        static let termsOfService = "https://www.naver.com/"
+        static let invitationThumbnail = "https://bagglebucket.s3.ap-northeast-2.amazonaws.com/link_participation.png"
+        static let kakaoAppStore = "itms-apps://itunes.apple.com/app/362057947"
+    }
+    
+    struct Account {
+        static let mail = "kingzzangssang@gmail.com"
+    }
+    
+    struct Image {
+        static let jpegCompression: CGFloat = 0.9
+    }
+    
+    struct ErrorID {
+        static let user: Int = -999
+        static let meeting: Int = -999
+    }
+}
+// swiftlint:enable all
