--- Baggle/Baggle/Core/Common/Const/ErrorID.swift
@@ -1,15 +1,0 @@
-//
-//  ErrorID.swift
-//  Baggle
-//
-//  Created by youtak on 2023/08/14.
-//
-
-import Foundation
-
-extension Const {
-    struct ErrorID {
-        static let user: Int = -999
-        static let meeting: Int = -999
-    }
-}
