--- Baggle/Baggle/Core/Common/Const/Image.swift
@@ -1,14 +1,0 @@
-//
-//  Image.swift
-//  Baggle
-//
-//  Created by youtak on 2023/08/17.
-//
-
-import Foundation
-
-extension Const {
-    struct Image {
-        static let jpegCompression: CGFloat = 0.9
-    }
-}
