--- Baggle/Baggle/Core/Common/Const/URL.swift
@@ -1,19 +1,0 @@
-//
-//  URL.swift
-//  Baggle
-//
-//  Created by 양수빈 on 2023/08/14.
-//
-
-import Foundation
-
-// swiftlint:disable all
-extension Const {
-    struct URL {
-        static let privacyPolicy = "https://www.dnd.ac/"
-        static let termsOfService = "https://www.naver.com/"
-        static let invitationThumbnail = "https://bagglebucket.s3.ap-northeast-2.amazonaws.com/link_participation.png"
-        static let kakaoAppStore = "itms-apps://itunes.apple.com/app/362057947"
-    }
-}
-// swiftlint:enable all
